requirements.txt
mlflow
mlflow_skinny
numpy
Pillow
Pillow
PyYAML
PyYAML
scikit_learn
torch
torchvision
tqdm
